Another big refactor
The point is to create two entry points into the app.

A DM View that has DM stuff and a Presentation View meant to be
shown to players. In the future, a Player View could be added to
allow players to also interact with the system. This structure also supports
a top-level authGuard for each "view". The presentation view would be public,
the DM view would only be viewable by Admin/DM's, and the PlayerView
would only be visible to Players.